vid: VidOrdering (Above|Level|Below) intead of cmp.
There are several different concepts when comparing variables in bex:
whether the variable number itself is higher or lower, and
whether the variable is placed higher or lower in a BDD/ANF/etc.

Also, we have distinctions between "real" and "virtual" variables,
as well as a constant called T that serves as the branch
variable for I and O, and also (at least for now) a special constant
used for nodes in the AST, where no variable annotation is
(currently) used. (Probably at some point, AST nodes will remember
the highest variable they depend on, and "VID::NoV" will go away.)

Anyway, I want to reverse the "industry standard" ordering of branch
variables in the graphs, putting nodes that branch on x0 (or whatever
the lowest-"numbered" variable is) at the bottom, and it was confusing
to search for every <, >, and cmp()  and decide whether I was really
comparing the index numbers or the depth.

So instead, we now have cmp_depth, is_below, is_above, etc.