solve: inline refine(), use decreasing vid rather than current top.
There is pretty much always a benefit to sorting the AST by cost,
and once you take that step, we pretty much know the "top" vid in
BDDBase/ANFBase. (It's possible the solution process cancels something
out and we send the solver a definition it has no use for, but this
will be quickly ignored by the actual BDD substitution logic.)

More importantly, the point of all this is to bring the swap solver
online, and the ctx nids returned by the swap solver aren't tied to
a variable. (Even if it were, the variable at the top of the scaffold
is not necessarily the next one to replace in the AST, since the whole
point of the swap solver is to constantly re-arrange the variables.)