make WorkState sharable via Arc<BddState>
wip::Work::Todo originally had a RefCell, but this
does not implement Sync, so we can't share it between
threads. For now, I just moved the work in progress
information back into the Work::Todo variant.

This makes the WorkCache entries much bigger than they need
to be, so I will come back to this again in the future.