vhl: avoid race condition when indexing new hilos.
The old code read the .len() of the hilos vector, and
then did a push, but I want to make appeding to the vector
thread-safe, and there's a chance another thread could
append something else before the reverse index was updated.

So, just use the return value from .push() as the index.